Add fetch funtions to apiCalls.js, working as expected
